+++ contracts/src/YansTokenUUPSv2.sol
@@ -9,7 +9,7 @@ import {YansTokenUUPS} from './YansTokenUUPS.sol';
  * @custom:oz-upgrades-from src/YansTokenUUPS.sol:YansTokenUUPS
  */
 contract YansTokenUUPSV2 is YansTokenUUPS {
-  function mint(address to, uint256 amount) external onlyOwner {
+  function mint(address to, uint256 amount) external virtual onlyOwner {
     _mint(to, amount);
   }
 }